Changing the structure
